[feat] Refactored to pass data between states on VerifyAccount
