+++ app/src/main/java/com/madteam/sunset/ui/screens/verifyaccount/state/VerifyAccountUIEvent.kt
@@ -1,0 +1,6 @@
+package com.madteam.sunset.ui.screens.verifyaccount.state
+
+sealed class VerifyAccountUIEvent {
+    data object OnResend : VerifyAccountUIEvent()
+    data class OnCheck(val credential: String) : VerifyAccountUIEvent()
+}
