+++ app/src/main/java/com/madteam/sunset/ui/screens/verifyaccount/state/VerifyAccountUIState.kt
@@ -1,0 +1,7 @@
+package com.madteam.sunset.ui.screens.verifyaccount.state
+
+data class VerifyAccountUIState(
+    val resendCounter: Int = 0,
+    val recheckCounter: Int = 0,
+    val isVerified: Boolean = false
+)
