+++ app/src/main/res/values/strings.xml
@@ -156,14 +156,16 @@
     <string name="created">"Created "</string>
     <string name="likes">likes</string>
     <string name="view_all">View all</string>
-    <string name="comments">comments</string>
-    <string name="be_the_first_to_comment">Be the first to comment</string>
-    <string name="review">Review</string>
-    <string name="reviewed_by">Reviewed by</string>
-    <string name="reported_by">Reported by</string>
-    <string name="assigned_by">Assigned by</string>
-    <string name="unnasigned_report">No assigned yet</string>
-    <string name="spotted_by">Spotted by</string>
-    <string name="kms_away">Kms away</string>
-    <string name="reviews">reviews</string>
+  <string name="comments">comments</string>
+  <string name="be_the_first_to_comment">Be the first to comment</string>
+  <string name="review">Review</string>
+  <string name="reviewed_by">Reviewed by</string>
+  <string name="reported_by">Reported by</string>
+  <string name="assigned_by">Assigned by</string>
+  <string name="unnasigned_report">No assigned yet</string>
+  <string name="spotted_by">Spotted by</string>
+  <string name="kms_away">Kms away</string>
+  <string name="reviews">reviews</string>
+  <string name="resend">Resend</string>
+  <string name="not_arriving">Not arriving?</string>
 </resources>
